requirements.txt
streamlit
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
pandas
numpy

